refactor: 피드백 관련 바로 고칠 수 있는 부분 리팩토링
