+++ src/main/java/lotto/domain/strategy/ManualStrategy.java
@@ -1,7 +1,7 @@
-package lotto.strategy;
+package lotto.domain.strategy;
 
-import lotto.domain.LottoTicket;
-import lotto.domain.LottoTickets;
+import lotto.domain.ticket.LottoTicket;
+import lotto.domain.ticket.LottoTickets;
 
 import java.util.ArrayList;
 import java.util.List;
