+++ src/main/java/lotto/domain/ticket/LottoNumber.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.ticket;
 
 import java.util.Objects;
