+++ src/main/java/lotto/domain/ticket/LottoTickets.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.ticket;
 
 import java.util.ArrayList;
 import java.util.Collections;
