+++ src/main/java/lotto/domain/ticketresult/LottoComparator.java
@@ -1,6 +1,6 @@
 package lotto.domain.ticketresult;
 
-import lotto.domain.LottoTicket;
+import lotto.domain.ticket.LottoTicket;
 import lotto.domain.ticketpurchase.PurchasedTickets;
 
 import java.util.Collections;
