+++ src/test/java/lotto/domain/LottoTicketTest.java
@@ -1,5 +1,7 @@
 package lotto.domain;
 
+import lotto.domain.ticket.LottoNumber;
+import lotto.domain.ticket.LottoTicket;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
