+++ src/test/java/lotto/domain/ticketgenerator/LottoGeneratorTest.java
@@ -1,10 +1,10 @@
 package lotto.domain.ticketgenerator;
 
-import lotto.domain.LottoNumber;
-import lotto.domain.LottoTicket;
-import lotto.domain.LottoTickets;
-import lotto.strategy.AutoStrategy;
-import lotto.strategy.ManualStrategy;
+import lotto.domain.ticket.LottoNumber;
+import lotto.domain.ticket.LottoTicket;
+import lotto.domain.ticket.LottoTickets;
+import lotto.domain.strategy.AutoStrategy;
+import lotto.domain.strategy.ManualStrategy;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
